+++ lib/helpers/code_generator_helper.rb
@@ -1,3 +1,6 @@
+require "google_drive"
+require "active_support/concern"
+
 module CodeGeneratorHelper
   extend ActiveSupport::Concern
